AI_Knowledge_Assistant/requirements.txt
openai
sentence-transformers
faiss-cpu
pandas
langchain
rich
pypdf
ragas
deepeval
datasets
